Refactor Network and DryRunNetworkWorker to share code
They do many things the same, so this refactors out those pieces. The main entrypoint of
both processes is called go.
